third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-ua-style.html

<!DOCTYPE html>
<script src="full-screen-test.js"></script>
<script>
    function runTest() {
        document.offsetWidth;
        frame.contentDocument.offsetWidth;
        var videoEnteredFullScreen = function() {
            test("getComputedStyle(ancestor).position == 'static'");
            endTest();
        };
        waitForEvent(frame.contentDocument, "webkitfullscreenchange", videoEnteredFullScreen);

        test("getComputedStyle(ancestor).position == 'relative'");

        var x = frame.getBoundingClientRect().left + 10;
        var y = frame.getBoundingClientRect().top + 10;
        eventSender.mouseMoveTo(x, y);
        eventSender.mouseDown();
        eventSender.mouseUp();
    }
</script>
<div id="ancestor" style="position:relative">
  <iframe allowfullscreen id="frame" src="resources/inner-video.html" onload="runTest()"></iframe>
</div>
