third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-ua-style-expected.txt


TEST(getComputedStyle(ancestor).position == 'relative') OK
EVENT(webkitfullscreenchange)
TEST(getComputedStyle(ancestor).position == 'static') OK
END OF TEST
